Interactions included
Included a self-made SHAP-Interaction Term based on chapter 9.6.8 in https://christophm.github.io/interpretable-ml-book/shap.html#shap-interaction-values
This seems to work for the simple example at the top. For the Lorenz-System, it manages to show that there are no interactions in dx_0/dt and that there is a x_0-x_2-interaction in dx_2/dt. But its results are inconclusive with regards to dx_1/dt.