Simulated Annealing + Temperature Scaling
Two similar concepts based on the physical role of temperature: Simulated annealing optimises functions with many local minima by starting in a "hot" environment that provides enough energy to jump out of local minima and by subsequently cooling down and closing in on the global minimum. Temperature scaling adjusts a Neural Net's confidence in its prediction (the output of the softmax function) by scaling it with an optimised temperature.